backend/requirements.txt
# Backend dependencies
flask==3.1.2
pandas==2.3.3
openpyxl==3.1.5

# Main application dependencies
requests
beautifulsoup4
lxml
PyPDF2
faiss-cpu
google-generativeai
python-dotenv
reportlab